13.01
essa fork será pra uma calculadora mais complexa.
adicionada a função de exponenciação, mas será necessario arrumar a leitura de novos calculos
